move fw performance model into a library
